More more terminal typeAhead code into terminalContrib
