+++ src/vs/workbench/contrib/terminalContrib/typeAhead/test/browser/terminalTypeAhead.test.ts
@@ -8,11 +8,12 @@ import type { IBuffer, Terminal } from '@xterm/xterm';
 import { SinonStub, stub, useFakeTimers } from 'sinon';
 import { Emitter } from 'vs/base/common/event';
 import { CharPredictState, IPrediction, PredictionStats, TypeAheadAddon } from 'vs/workbench/contrib/terminalContrib/typeAhead/browser/terminalTypeAheadAddon';
-import { DEFAULT_LOCAL_ECHO_EXCLUDE, IBeforeProcessDataEvent, ITerminalConfiguration, ITerminalProcessManager } from 'vs/workbench/contrib/terminal/common/terminal';
+import { IBeforeProcessDataEvent, ITerminalProcessManager } from 'vs/workbench/contrib/terminal/common/terminal';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { TestConfigurationService } from 'vs/platform/configuration/test/common/testConfigurationService';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
 import { DisposableStore } from 'vs/base/common/lifecycle';
+import { DEFAULT_LOCAL_ECHO_EXCLUDE, type ITerminalTypeAheadConfiguration } from 'vs/workbench/contrib/terminalContrib/typeAhead/common/terminalTypeAheadConfiguration';
 
 const CSI = `\x1b[`;
 
@@ -84,7 +85,7 @@ suite('Workbench - Terminal Typeahead', () => {
 	suite('timeline', () => {
 		let onBeforeProcessData: Emitter<IBeforeProcessDataEvent>;
 		let publicLog: SinonStub;
-		let config: ITerminalConfiguration;
+		let config: ITerminalTypeAheadConfiguration;
 		let addon: TestTypeAheadAddon;
 
 		const predictedHelloo = [
@@ -103,7 +104,7 @@ suite('Workbench - Terminal Typeahead', () => {
 
 		setup(() => {
 			onBeforeProcessData = ds.add(new Emitter<IBeforeProcessDataEvent>());
-			config = upcastPartial<ITerminalConfiguration>({
+			config = upcastPartial<ITerminalTypeAheadConfiguration>({
 				localEchoStyle: 'italic',
 				localEchoLatencyThreshold: 0,
 				localEchoExcludePrograms: DEFAULT_LOCAL_ECHO_EXCLUDE,
